+++ pyoxidizer/src/starlark/python_resource.rs
@@ -11,24 +11,84 @@ use {
         PythonPackageResource as RawPackageResource, PythonResource,
     },
     starlark::environment::Environment,
-    starlark::values::{default_compare, RuntimeError, TypedValue, Value, ValueError, ValueResult},
+    starlark::values::{
+        default_compare, RuntimeError, TypedValue, Value, ValueError, ValueResult,
+        INCORRECT_PARAMETER_TYPE_ERROR_CODE,
+    },
     starlark::{any, immutable, not_supported},
     std::any::Any,
     std::cmp::Ordering,
     std::collections::HashMap,
+    std::convert::{TryFrom, TryInto},
 };
+
+/// Where a resource should be loaded from.
+#[derive(Clone, Debug)]
+pub enum ResourceLocation {
+    /// Use default load semantics for the target binary.
+    Default,
+    /// Load the resource from memory.
+    InMemory,
+    /// Load the resource from a filesystem path relative to the binary.
+    RelativePath(String),
+}
+
+impl From<ResourceLocation> for Value {
+    fn from(location: ResourceLocation) -> Self {
+        Value::new(match location {
+            ResourceLocation::Default => "default".to_string(),
+            ResourceLocation::InMemory => "in-memory".to_string(),
+            ResourceLocation::RelativePath(prefix) => format!("filesystem-relative:{}", prefix),
+        })
+    }
+}
+
+impl TryFrom<Value> for ResourceLocation {
+    type Error = ValueError;
+
+    fn try_from(value: Value) -> Result<Self, Self::Error> {
+        let s = value.to_str();
+
+        if s == "default" {
+            Ok(ResourceLocation::Default)
+        } else if s == "in-memory" {
+            Ok(ResourceLocation::InMemory)
+        } else if s.starts_with("filesystem-relative:") {
+            let prefix = s.split_at("filesystem-relative:".len()).1;
+            Ok(ResourceLocation::RelativePath(prefix.to_string()))
+        } else {
+            Err(RuntimeError {
+                code: INCORRECT_PARAMETER_TYPE_ERROR_CODE,
+                message: format!("unable to convert value {} to a resource location", s),
+                label: format!(
+                    "expected `default`, `in-memory`, or `filesystem-relative:*`; got {}",
+                    s
+                ),
+            }
+            .into())
+        }
+    }
+}
 
 #[derive(Debug, Clone)]
 pub struct PythonSourceModule {
     pub module: RawSourceModule,
+    pub location: ResourceLocation,
+}
+
+impl PythonSourceModule {
+    pub fn new(module: RawSourceModule) -> Self {
+        Self {
+            module,
+            location: ResourceLocation::Default,
+        }
+    }
 }
 
 impl TypedValue for PythonSourceModule {
     immutable!();
     any!();
-    not_supported!(
-        binop, dir_attr, function, get_hash, indexable, iterable, sequence, set_attr, to_int
-    );
+    not_supported!(binop, dir_attr, function, get_hash, indexable, iterable, sequence, to_int);
 
     fn to_str(&self) -> String {
         format!("PythonSourceModule<name={}>", self.module.name)
@@ -75,6 +135,7 @@ impl TypedValue for PythonSourceModule {
                 Value::new(source)
             }
             "is_package" => Value::new(self.module.is_package),
+            "location" => self.location.clone().into(),
             attr => {
                 return Err(ValueError::OperationNotSupported {
                     op: format!(".{}", attr),
@@ -90,11 +151,26 @@ impl TypedValue for PythonSourceModule {
     fn has_attr(&self, attribute: &str) -> Result<bool, ValueError> {
         Ok(match attribute {
             "name" => true,
-            // TODO expose source
-            // "source" => true,
+            "source" => true,
             "is_package" => true,
+            "location" => true,
             _ => false,
         })
+    }
+
+    fn set_attr(&mut self, attribute: &str, value: Value) -> Result<(), ValueError> {
+        match attribute {
+            "location" => {
+                self.location = value.try_into()?;
+
+                Ok(())
+            }
+            _ => Err(ValueError::OperationNotSupported {
+                op: format!(".{} =", attribute),
+                left: self.get_type().to_owned(),
+                right: None,
+            }),
+        }
     }
 }
 
@@ -372,7 +448,7 @@ impl TypedValue for PythonExtensionModule {
 
 pub fn python_resource_to_value(resource: &PythonResource) -> Value {
     match resource {
-        PythonResource::ModuleSource(sm) => Value::new(PythonSourceModule { module: sm.clone() }),
+        PythonResource::ModuleSource(sm) => Value::new(PythonSourceModule::new(sm.clone())),
 
         PythonResource::ModuleBytecodeRequest(m) => {
             Value::new(PythonBytecodeModule { module: m.clone() })
@@ -405,5 +481,48 @@ pub fn python_resource_to_value(resource: &PythonResource) -> Value {
         PythonResource::PathExtension(_) => {
             panic!("path extensions not supported");
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::super::testutil::*;
+    use super::*;
+
+    #[test]
+    fn test_source_module_attrs() {
+        let mut env = starlark_make_exe().unwrap();
+
+        let mut m = starlark_eval_in_env(
+            &mut env,
+            "exe.make_python_source_module('foo', 'import bar')",
+        )
+        .unwrap();
+
+        assert_eq!(m.get_type(), "PythonSourceModule");
+        assert!(m.has_attr("name").unwrap());
+        assert_eq!(m.get_attr("name").unwrap().to_str(), "foo");
+
+        assert!(m.has_attr("source").unwrap());
+        assert_eq!(m.get_attr("source").unwrap().to_str(), "import bar");
+
+        assert!(m.has_attr("is_package").unwrap());
+        assert_eq!(m.get_attr("is_package").unwrap().to_bool(), false);
+
+        assert!(m.has_attr("location").unwrap());
+        assert_eq!(m.get_attr("location").unwrap().to_str(), "default");
+
+        m.set_attr("location", Value::from("in-memory")).unwrap();
+        assert_eq!(m.get_attr("location").unwrap().to_str(), "in-memory");
+
+        m.set_attr("location", Value::from("default")).unwrap();
+        assert_eq!(m.get_attr("location").unwrap().to_str(), "default");
+
+        m.set_attr("location", Value::from("filesystem-relative:lib"))
+            .unwrap();
+        assert_eq!(
+            m.get_attr("location").unwrap().to_str(),
+            "filesystem-relative:lib"
+        );
     }
 }